backend/requirements.txt
fastapi
uvicorn
opencv-python
pydantic
python-multipart
pyyaml
rapidfuzz
jinja2
openai
